+++ src/components/Field/FormFieldAddress/Address.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import { boolean, number, text, withKnobs } from "@storybook/addon-knobs";
 import { renderButtons } from "@root/utils/storyUtils";
-import { FieldDef } from "../../components/Field";
+import { FieldDef } from "..";
 import Form, { useForm } from "@root/components/Form";
 import { getOptionsCountries, getOptionsStates } from "./utils/optionGetters";
