+++ src/components/Field/FormFieldAddress/utils/optionGetters.ts
@@ -1,5 +1,5 @@
 import { MosaicLabelValue } from "@root/types";
-import countriesWithStates from "@root/forms/FormFieldAddress/utils/trimmedCountriesStates.json";
+import countriesWithStates from "@root/components/Field/FormFieldAddress/utils/trimmedCountriesStates.json";
 
 export const getOptionsCountries = async (): Promise<MosaicLabelValue[]> => {
 	return Promise.resolve(countriesWithStates.map((country) => ({
