+++ src/components/Field/index.tsx
@@ -1,3 +1,6 @@
 export { default } from "./Field";
 export * from "./Field";
 export * from "./FieldTypes";
+
+export { default as FormFieldAddress } from "./FormFieldAddress";
+export * from "./FormFieldAddress";
